+++ src/VitessceGrid.test.js
@@ -1,9 +1,15 @@
 import expect from 'expect';
-import { resolveLayout } from './VitessceGrid';
+import React from 'react';
+import { shallow, render, configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
 
-describe('VitessceGrid', () => {
+import { resolveLayout, VitessceGrid } from './VitessceGrid';
+
+configure({ adapter: new Adapter() });
+
+describe('VitessceGrid.js', () => {
   describe('resolveLayout', () => {
-    const layout = [
+    const componentsSpec = [
       { component: 'NoProps', x: 0, y: 0 },
       {
         component: 'HasProps', props: { foo: 'bar' }, x: 1, y: 1, w: 1, h: 1,
@@ -30,7 +36,7 @@ describe('VitessceGrid', () => {
           1000: [0, 3, 9, 12],
           800: [0, 4, 8, 12],
         },
-        layout,
+        components: componentsSpec,
       });
       expect(cols).toEqual({ 800: 12, 1000: 12 });
       expect(layouts).toEqual(
@@ -64,7 +70,7 @@ describe('VitessceGrid', () => {
       const {
         cols, layouts, breakpoints, components,
       } = resolveLayout(
-        layout,
+        componentsSpec,
       );
       expect(cols).toEqual({ ID: 12 });
       expect(layouts).toEqual(
@@ -81,6 +87,60 @@ describe('VitessceGrid', () => {
       );
       expect(breakpoints).toEqual({ ID: 1000 });
       expect(components).toEqual(expectedComponents);
+    });
+  });
+
+  describe('<VitessceGrid />', () => {
+    function FakeComponent(props) {
+      const { text } = props;
+      return <span>{text}</span>;
+    }
+    /* eslint-disable object-curly-newline */
+    /* eslint-disable object-property-newline */
+    const layoutJson = {
+      columns: {
+        600: [0, 2, 4, 8],
+      },
+      components: [
+        { component: 'FakeComponent',
+          props: { text: 'Hello World' },
+          x: 0, y: 0, w: 2 },
+      ],
+    };
+    /* eslint-enable */
+    it('shallow() works', () => {
+      const wrapper = shallow(<VitessceGrid
+        layout={layoutJson}
+        getComponent={() => FakeComponent}
+        draggableHandle=".my-handle"
+      />);
+
+      expect(wrapper.find('div').length).toEqual(1);
+      expect(wrapper.find('div').text()).toEqual('<FakeComponent />');
+
+      expect(wrapper.find('span').length).toEqual(0);
+
+      const style = wrapper.find('style');
+      expect(style.length).toEqual(1);
+      expect(style.text()).toContain('.my-handle {');
+      expect(style.text()).toContain('.my-handle:active {');
+    });
+
+    it('render() works', () => {
+      const wrapper = render(<VitessceGrid
+        layout={layoutJson}
+        getComponent={() => FakeComponent}
+        draggableHandle=".my-handle"
+      />);
+
+      expect(wrapper.find('div').length).toEqual(1);
+      expect(wrapper.find('div').text()).toEqual('Hello World');
+
+      expect(wrapper.find('span').length).toEqual(2);
+
+      const style = wrapper.find('style');
+      expect(style.length).toEqual(0);
+      // TODO: Why does render() not generate style?
     });
   });
 });