+++ src/index.js
@@ -1,5 +1,3 @@
 import { VitessceGrid } from './VitessceGrid';
 
-export default {
-  VitessceGrid,
-};
+export default VitessceGrid;