+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
-## [0.0.?] - 2019-05-13 - in progress
+## [0.0.?] - 2019-05-15
 ### Added
-- Created demo
+- Testing
+- Pass properties through to react-grid-layout.
+- Lots of comments in demo to explain features.
+
+## [0.0.1] - 2019-05-13
+### Added
+- Port this reusable code from Vitessce.